requirements.txt
Django==2.2.24
django-braces==1.14.0
django-crispy-forms==1.12.0
Pillow==8.3.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
